test with mocks and spring boot - done
