+++ homework3/src/test/resources/testquiz_en.csv
@@ -1,5 +1,5 @@
 #comments here
-What is the capital of the Great Britan?, London, Delhi, New York, Washington
+Are we just part of testing?, yes
 #more comments
 incorrect data
 #
